Corrige le taux de TVA réduit 5,5 % (devis/prestations/paramètres)
Audit de l'app electricien-devis (MELLADO Électricité) : le taux de TVA
5,5 % — annoncé dans le README et proposé dans les sélecteurs de
l'interface (ligne de devis, catalogue de prestations, TVA par défaut
de l'entreprise) — ne pouvait en réalité jamais être enregistré.

Cause : les colonnes `tauxTva` / `tauxTvaDefaut` étaient des entiers
(Int) en base, et les validations Zod des server actions imposaient
`.int()`. Choisir 5,5 % puis enregistrer renvoyait silencieusement
« Données du devis invalides. » (devis) ou « Formulaire invalide. »
(prestation, paramètres) : impossible d'appliquer ce taux, pourtant le
taux réduit réel pour les travaux de rénovation énergétique — un cas
d'usage courant pour un électricien.

Correctifs :
- schema.prisma : `tauxTva`/`tauxTvaDefaut` passent de Int à Float
  (migration 3_taux_tva_decimal ajoutée et intégrée au script de démo)
- Zod : retrait de `.int()` sur ces trois champs (bornes 0-20 conservées)
- Champ « TVA par défaut » (paramètres) : step 1 → 0,5 pour permettre
  la saisie de 5,5 au clavier/stepper
- Ajout de `formatTaux()` (notation FR, virgule) et remplacement des
  affichages bruts `{taux} %` dans l'éditeur de devis, les PDF devis/
  facture, l'email de devis et le catalogue de prestations, pour éviter
  d'afficher "5.5 %" (point) après le correctif

Vérifié : build/lint/typecheck OK, et test d'intégration contre une
base Postgres locale migrée (insertion + relecture d'une ligne à
tauxTva=5.5, valeur conservée exactement).

+++ electricien-devis/src/app/tableau-de-bord/devis/actions.ts
@@ -55,7 +55,7 @@ const ligneSchema = z.object({
   quantiteMilli: z.number().int().min(0),
   unite: z.string().trim().min(1),
   prixUnitaireCents: z.number().int(),
-  tauxTva: z.number().int().min(0).max(20),
+  tauxTva: z.number().min(0).max(20),
 });
 
 const payloadSchema = z.object({
