+++ electricien-devis/src/app/tableau-de-bord/parametres/actions.ts
@@ -18,7 +18,7 @@ const schema = z.object({
   assurance: z.string().trim(),
   iban: z.string().trim(),
   prefixeDevis: z.string().trim().min(1, "Le préfixe est obligatoire.").max(10),
-  tauxTvaDefaut: z.coerce.number().int().min(0).max(20),
+  tauxTvaDefaut: z.coerce.number().min(0).max(20),
   dureeValidite: z.coerce.number().int().min(1).max(365),
   relanceJours: z.coerce.number().int().min(1).max(90),
   mentionsLegales: z.string().trim(),
